+++ store/src/app/products/button/button.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { CartService } from '../../cart/cart.service';
+import { CartService } from '../cart/cart.service';
 
 @Component({
   selector: 'app-button',
