+++ store/src/app/products/products.module.ts
@@ -3,12 +3,12 @@ import { CommonModule } from '@angular/common';
 import { HeaderComponent } from './header/header.component';
 import { ProductCardComponent } from './product-card/product-card.component';
 import { ProductInfoComponent } from './product-info/product-info.component';
-import { ButtonComponent } from './product-card/button/button.component';
+import { ButtonComponent } from './button/button.component';
 import { FooterComponent } from './footer/footer.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { RouterModule } from '@angular/router';
-import { HryvniaPipePipe } from './hryvnia-pipe.pipe';
-import { ColorPriceDirective } from './color-price.directive';
+import { HryvniaPipePipe } from '../pipes/hryvnia-pipe.pipe';
+import { ColorPriceDirective } from '../directives/color-price.directive';
 import { ProductsComponent } from './products.component';
 import { ProductsListComponent } from './products-list/products-list.component';
 import { CartComponent } from './cart/cart.component';
